backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2
django-colorfield==0.9.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework
djoser
drf-extra-fields==3.4.0
flake8==5.0.4
gunicorn==20.0.4
idna==3.3
isort==5.12.0
itypes==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
Pillow==9.1.1
psycopg2-binary==2.9.7
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
python-dotenv
python3-openid==3.2.0
pytz==2022.7
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13